requirements.txt
line-bot-sdk
flask==0.12.2
numpy
jieba
keras==2.2.4
tensorflow